requirements.txt
fastapi
uvicorn
pydantic
pyjwt
python-dotenv
google-generativeai
streamlit
requests